got impatient.  decided to add in some autoloaders.  I know that it won't apply to singleton instances that are in some of the directory structures setup but I have the structure there anyway.
I also anticipate that in the future this may change somewhat or be relocated or restructured but it should work as is for now.

Why did I do this?  b/c with the messages subsystem there are a LOT of intermixing usages of classes in different places and I don't want to think about having to include the files...
[touch: 1351]