remove existing tab indexes add `tabindex= "-1"` for buttons we want to skip tabs in the beginning.  this makes sure that the next tab after the title is add media and then the content.
This was just a quick fix that makes things a little better but not ideal.

The problem is, that tinymce is loaded pretty much at the end of the page and although it appears it can be indicated as a dependency in the wp_enqueue_script() dependency array, it in fact **cannot** (as per the actual code in the wp_core).  Also, although one of the possible parameters for wp_editor settings array is "tabindex", it adds the tabindex value there to the generated text area but after tinymce is initialized that text area is of course hidden and the new tinymce editor does NOT clone the tab index value (why I dont' know)

Anyway, long story short.  The solution I've put in place will have to do for now until:

a. We submit a patch to wp-core so tab index actually gets replicated to generated wp-editors or
b. we can use the tincymce key in the depends array to load our own javascript handler for this (if necessary).
c. We create a tinymce plugin to hook into the wp-tinymce instance (using the provided filters) to clone the tab index value.