modified the `EE_Parse_Shortcodes` helper so it utilizes the new EE_Shortcodes library files [touch: 2170]
When a message object is assembled,  the templates will still get sent to the EE_Parse_Shortcodes helper, however, the helper now also receives the $_valid_shortcodes properties for messenger and message type and uses that to instantiate the correct shortcode libraries and validate and parse the short codes.

The new method of parsing is much more efficient because we will only manipulate data for short codes in the template.  This also allows for easy modification/addition of new short codes.