modify contexts so they can be defined easier [touch: 2191]
Contexts are one of the powerful features of the messages system that provide a lot of flexibility for having different templates defined for different "contexts" in which the templates get used.

For example:
- Payment message types might have different recipients for outgoing payment messages (admin, attendee, primary_attendee).
- Newsletter message types might have different "themes" for outgoing newsletters (i.e. Discount newsletter, Haloween Newsletter) etc.

For now contexts are hardcoded with the message types but at some point in the future they could also become dynamic and a message type might allow defining new contexts in the message type settings (accessible via the messages "activation" page).

The changes in this commit are to make the ui more friendly for users as the "idea" of contexts can be confusing.

So now, Message Types:
- define a general context label (i.e. "Recipients") which is used whenever contexts are referred to in the UI.
- define details for each context in the _contexts property.  This array has the context slug as the key and an associative array with the label and description of the context.