removed "add new message template" button/path from messages overview page. removed intermediary step for custom event template creation when there is no existing custom event template available when switching templates for an event. [completed: 2336]
The only thing that needs to be done from the overview page is editing any existing templates.  When in the event editor, users will have the option to add a custom event template for different messenger/message_type combinations.  I've streamlined the process so users are no longer given a drop down form if the event doesn't already have an existing custom template.  Instead, since we already know the messenger/message type combination, we'll just go ahead and generate a new custom template for that event (for the given messenger/message_type combination) using the global templates as the base for it.  Users only need to modify the contexts they want to modify for that event.