active_message_types are not relevant in most cases.  Instead we need installed message types.  I've changes all the methods and references to reflect this. [touch: 2374]
Why?

For our purposes.  The only thing we need to know if is active or not is the messenger.  That is the trigger to determine whether something get's "sent" or not.  IF the messenger is active, then we only SEND things for a messenger that has that message type active with it.