repositioned where our admin_Init_global hook and the _current_page check were in the page_setup method.  We don't need EE_Admin pages loading so much stuff on every page load. [touch: 2541]
The primary need for the position of the admin_init hook is so that EE_Admin pages can still hook into other WordPress pages if necessary.  However, keep in mind that the location of this hook means that a number of things in the EE_Admin object are still not setup.