add in new feature for EE Models so that you can set a specific limit per table in multi-table models [touch: 2476]
There are cases where you want to do paging but it becomes different across multiple tables especially in one to many relationships.  However, it is possible using a query structure such as:

    SELECT t1.something, t2.something FROM (SELECT * FROM table_one LIMIT 1,1) AS t1 LEFT JOIN table_two as t2 ON t1.primary_key = t2.foreign_key

This query will do the paging on table_one and then join THOSE results with the second table.  Thus, even if the second table join returns multiple rows, the rows returned are constrained by the limit set on table one.

So what I've done is add a new query_param called 'on_join_limit'.  Basically it allows you to set the table you want to do the specific limit on in the format:

    array('on_join_limit' => array('table_one', array(1,1) ) );

Then I made modifications to the EEM_base and EEM_secondary and EEM_Primary classes so that if the 'on_join_limit' query_param is detected it will setup the join_sql accordingly.  NOTE, this should not affect any existing models because its' ONLY kicked off when the on_join_limit param is present.