modified EE_Registry _load() method [touch: 3242]
##FIRST:  EE_Base_Class objects have two ways of instantiating:
1. new_instance (when instantiated by client code)
2. new_instance_from_db (when instantiated by the models when creating an array of EE_Base_Class objects)

So I modified the EE_Registry to check for those two methods and added a simple flag "$from_db" that the model uses to trigger using the "new_instance_from_db" method.

##SECOND
When passing an array of arguments to the db we should be using "call_user_func_array" NOT call_user_func.  Otherwise we'll have to do a refactor of all the class constructors to only have one argument… not a good idea.  Much simpler to use call_user_fun_array() :)