add a skeleton file for working out the new ui for the date times and tickets in the event editor. [touch: 3303]
I'm doing this is so I can work out the form structure, js, and styling before breaking out the different components into files and substituting our generated template vars for the hard-coded values.

Note: I also decided to go with the "data-" attribute provided by html5 spec (which is backwards compatible) as this is something we can use with jQuery to easily reference different click events and the other related items for that click event.  So how I'm using it is this:

1. Anything "clickable" by the user stores the reference for the related id in the "data-" attribute.
2. All "targets" can be easily calculated from the "data-" attributes on the clicked object.

For example.
The "gear" icon for editing a datetime is a clickable item.  So I have two attributes on the span used for the icon:
`data-datetime-row` = the row number for the datetime being edited.
`data-context` = the context for the edit event (i.e. 'datetime' or 'ticket').

So when the user kicks off an event by clicking a ".gear-icon" element, then the jquery will grab the "datetimeRow" value and the "context" value and use that to generate the correct id for the edited datetime which is "edit-event-datetime-row1" (if the datetimeRow value=1).

The above illustrates what I'm attempting to do.  However, note that I'll probably tweak it a bit.