removing obsolete code and setting up structure for implementation of caffeinated Event Datetime and Ticket metabox [touch: 3303]
Note: In this iteration I'm putting this in the `espresso_events_Pricing_Hooks` file to save time over setting up a new Ticketing Admin page route and putting everything in there.  However, the final destination of this code WILL be in `espresso_events_Ticket_Hooks` and I'll be naming any actions/filters I put in here accordingly so when I DO create the new admin page for Tickets I won't have to rename.