fix minutes not sticking when triggering date/time picker attempt to fix greyed out date times bug. (see notes below) [touch: 3790]
Okay the bug
1. Click datetim start date
2. click datetime end date
3. edit a ticket and click the ticket start time.  Whatever date is there gets switched to the datetime end date but the TIME stays correct.

If you go this way:
1. reload page.
2. click new datetime field for start date.
3. click new datetime field for end date.
4. click start date for a ticket.
5. this time the DATE will stay correct but the time changes.

however....
1. reload page
2. click anything BUT the start date for the datetime
3. click another datetime field
4. click the start date of a ticket.
5. everything works as expected and the bug cannot be reproduced.

The only way I was able to *fix* this was to comment out the setting of min/max values on the datetimepicker() method (via our object).  So for now, thats the fix I'm rolling with and we'll have to revisit on a later date.