add help tour framework to EE_Admin_Page so that child pages can easily setup help tours. [touch: 3341]
Note: I've tried to make this pluggable as possible (via filters).  So its possible for add ons to inject their own stops into a already existing help tour.  However, it's probably easiest in most cases for add ons to just generate their own help tours.

The filters do make it possible though, for the number and placement of stops to be adjusted between decaf and caf functionality (and via the EE_Admin_Hooks system, which is still to be implemented).