fix for nonexistent object error on ee registration list table [completed: 3958]
Originally I had thought that I'd just do an is_admin() check in EE_Base_Class get_first_related() and get_many_related() and override the default where conditions there.  Realized that wouldn't work and then thought about just overriding the default_where_conditions in EEM_Base::get_all_related() method as it is the root for all get_related type queries.

However, in the end I decided that I think its better to leave the default_where_conditions in place inside the model and model objects and just allow client code to override as necessary.  Otherwise it means going through anywhere the models are called and refactoring client code.

So in the end the solution was to just modify how the registration list table is getting the ticket and date times making the assumption that there may be archived (deleted) tickets and/or date times attached to a registration.

I think in the end default_where_conditions should only be set for important things like post_type where we HAVE to have that on every query.  Not for more transient things like whether an item is deleted, or active, or a certain status.  In those cases I think client code should be required to be explicit about what it wants.  But refactoring things at this point is maybe better left for a later iteration (if done at all).