Add a check for when the thumbnail height is 0 and add some height to the img_height [touch: 3946]
This is for when the calendar loads up initially. For some reason the thumbnail height returns 0. So for that initial load, we'll add some height instead of removing it. The downside is landscape oriented thumbnails will have a little more padding at the bottom of the event on that initial load.