oooo forgot I had to make the question related short codes caffeinated only.  So this entailed adding a bunch of filters, move around files and what not so that we can make it work this way [touch: 2392]
Note:  I'm not doing any filtering in the data handlers as I think they should deal with any information that is available in decaf.  So the data handlers and EE_Messages_Addressee objects will still have all the setup objects.  However all the implementation of short codes using that data will be segregated for whether its decaf or regular functionality.