fix issue with prepare_for_pretty_echoing on EE_Money_Field [touch: 4315]
What was happening is that number_format get's run first when setting up $pretty_float,  and then when sent to EEH_template::format_currency, a float() typecast was don on the incoming format which was sent as $pretty_float.  Thus, if $pretty_float was something like 1,000.000 - it got changed to 1.00.