add $this_event_id back in [touch: 4343]
I'm not sure why this was removed. We noticed that in testing that if the Multi Event Registration cart link was in an event's description you'd end up with errors when you view the calendar tooltips. With this back in, the cart link actually works (you can add events to the cart from a calendar tooltip).