add a "show_once_only" option to the qtips helper. [touch: 4493]
The purpose of this feature is so that in configuration, the qtip can be setup to only show ONCE and then never again for the current browser (via setting a cookie).  This is useful for tips that need to appear the first time a feature is used but not thereafter.