fix event display for small screens [touch: 4595]
Events with an image would still be displayed because the JS was setting a height. Events that did not have an image would disappear on smaller screens when the title and time block's display was set to display: none;

Add max-height and min-height for the smallest media query so these display and a few other minor CSS tweaks to make the events look better on small displays.