modify the EE_Event->perform_sold_out_status_check() so it considers tickets attached to date times as well when validating whether an event has sold out or not. - Also tweak is_sold_out() method so it can be used to perform_sold_out_status_check() and return the ACTUAL sold out status rather than the SET event status (which may be different). [touch: 4990] [touch: 4705]
Note: I also added some commenting about some possible future considerations regarding flipping the sold status FROM sold out to some other post_status if this check reveals the event is NO LONGER sold out (since it's performed when event is edited as well).

For now, I just always return the actual sold out state as a bool.  This means that sometimes the Event status may remain at "sold_out" but in fact NOT be sold out because of the dtt calcs.  So this method CAN be used to obtain the actual status if necessary.