redirect to EE About page after new activation, reactivation, and upgrade [touch: 5071]
Note: I think reactivation status only happens when EE is reactivated with a new version but there is no db upgrade, which is the same as the upgrade status except in that case there IS a db upgrade.

So just deactivating and reactivating EE without a version bump will NOT redirect to about page (at least from my testing).