MTP_is_override now plays a more critical role with the new custom template system [touch: 5476]
Basically flagging a global template to override all custom templates will do the following:

1. Remove any ui for creating new custom templates based off of existing global templates. (so create buttons disappear from default template list table and the event editor metabox.
2. any existing custom template for a message_type/messenger that has a global with override set will not appear as a selectable option in the messages metabox in the event editor (even if the event was already assigned to it)... this means on save that the Event will be reassigned to the global template.
3. When messages are generated, global templates with override set will, um, OVERRIDE any pre-existing custom templates that may be assigned to the event.

Bottom line, override provides EM's with a method for enforcing a certain global template to be used for all events (they may want to do this for certain payment templates etc).  I imagine when we integrate more granular capabilities that the ability to change the override setting (or edit global templates even) will be a certain capability level (actually I'm going to add that to the capability spreadsheet).