okay a tweak to the solution.  What we do is on __sleep we save EACH addon set in the add ons property to its own wp_options table.  Then on __wakeup we retrieve that.
The ONLY caveat to this setup, is that if EE_Config is serialized for any other reason other than when its saved to the db, the add ons  property set at the time of serialization is persisted in the db.  I think this is an okay tradeoff tho cause why the heck would we be serializing EE_Config except for when its saved to the db?
[touch: 5815]