tweak test factory for EE_Event add test factory for EE_Datetimes
Note that I also introduced a special factory property naming schema for items that we want to chain to the parent related item.
Anything with a "_chained" after the property name means that it is to be created with the correct relation setup to its parent related items.

So in the case of datetime.  If a test case goes $this->factory->datetime_chained->create(), the factory will generate a single datetime, but also generate a related event that the datetime will then get assigned to.

If test case does something like $this->datetime_chained->create_many(2), then two date times would get created and attached to ONE event.

This is just intended as a quick way of generated the EE_Base_Class object with any PARENT relations setup.

The chaining is one way and should always start with the lowest link in the chain.  Here's a rough idea of the hierarchy:  Transaction->Registration->Ticket->Datetime->Event.