fix bug introduced by WP4.0 new "feature" that breaks if edit-form-advanced.php is included on a NON official wp route. [touch: 6239]
This new feature wp introduced has the tinymce editor follow the user as they scroll (collapsing) so that the toolbar is always accessible.  However it had hardcoded dependency on it being loaded on an official wp post/page (or cpt) route.   In other words, its not a feature exposed for anything OTHER than the advanced-form-edit page.