remove print button from default templates for receipt and invoice [touch: 6007]
Why?  Because WP strips any javascript in the content.  I don't think this is a huge deal, but if a print button is absolutely needed, at a later date we can add some html for the button but then have the html messenger enqueue some default javascript for the print button js.