fix typo in property name and everywhere this was referenced [touch: 6301]
Gotta say though... at least whoever put this typo in was consistent lol.

Oh and I realize this doesn't currently appear to be a property that is used anywhere but I assume it will likely be used for selecting whether events show in the dashboard or not so I left in place for now.