make sure new mtg has all existing relations removed on message templates so that it doesn't overwrite existing relations [touch: 6373]
This fixes the bug reported in the ticket and it appears to be a but with EE_Base_Class... however I can't seem to reproduce with a test.  I'm including the test I created in this commit to try to reproduce what I was experiencing in this case but it passes!  So I guess its a flukey bug.