remove the extraneous "all" view from message template list table [touch: 6430]
The "all" view was intended to show all message templates including inactive ones.  However the purpose of that is kind of silly since if we try to get inactive templates from the db for a message type or messenger that does not exist... then ERROR ERROR.  So its better to just ONLY ever show templates that match ACTIVE messengers or message types which means the view would be the same as "in use".