implement another param to can_checkin to flag whether to include reg status in verification [touch: 6631]
Note, I also went through where I had implemented can_checkin in existing methods, and ensure that default behaviour for those methods follows current behaviour in the admin.  i.e. if an admin triggers a checkin for a registration admin side, then it will be forced regardless of reg_status for that registration.