Added echo/return capabilities to template tags
Was driving me mad that there were no easy way to get the RIGHT timezone datetimes. So this adds the ability to return the values of many of the template tag functions which previously only echoed results. Has been set as branch of master as per https://github.com/eventespresso/event-espresso-core/pull/38