remove obsolete and totally unnecessary conditionals [touch: 7475]
This conditional has caused two breakages that I know of.

The purpose for this conditional is really murky and doesn't fit with the current code flow.  Parsers return either an empty string or a parsed value (or a unparsed shortcode for invalid shortcode).  It's likely this code existed from when I first wrote the parser and had false returned as a possible value.