change EE_Base_Class::fix_date_format-for_use_with_strtotime() to EE_Base_Class->get_strtotime() [touch: 7056]
The purpose of this method is to get the strtotime value of a date string while accounting for possible fixes needed in date format, without changing the possibly cached formats on the date time fields.

This allows us to still correct for ambigious formatted time strings being converted via strtotime while still able to use those formats for the outputted date string.