fixes to account for different incoming data formats for parsing [touch: 7056]
In the old system, a null date_string was handled by the php date methods used.  But in the new system we need to make sure that a value is set but we ONLY do so if the field is not nullable.  So in the case of a null timestring value, we set it at a default of current_time('mysql', true) and set the timezone to UTC.

Also in this commit, we add handling for when the incoming date string is in mysql format and ensure that the format_string matches that.  Prior to the changes in this branch, that was handled automatically by the date methods we use, however now that we are exclusively using createFromFormat, we need to make sure the format is set correctly.  Mysql is used when the db is setting the format internally on a related model query.