fix bug with not accounting for current_time('timestamp') having the offset applied to the unixtimestamp when creating a DateTime object from unixtimestamp [touch: 7056]
In order to protect backward compat, we need to make sure that when an incoming unix-timestamp is being set on a EE_Date_Time field, we consider the same assumption currently in client code which is that the unixtimestamp has the offset applied for the timezone.  Since DateTime object considers all unix timestamps incoming to be with no offset, we need to instantiate it using a date_string and timezone.