consistentivize _timezone string on model from model_field [touch: 7056]
When EEM_Base::get_formats_for() is called.  The code makes sure the internal timezone property matches what is set in EE_Datetime_Field.  This fixed some unit tests that were failing because it was possible for EEM_Base::$_timezone to get out of sync with what was set internally in EE_Datetime_Field(s).