Bugfix for [ESPRESSO_EVENTS order_by="venue_title"]
Sorting the event list by venue_title doesn not work (tested with 4.4.5 but seems to also be the case in master branch). The shortcode [ESPRESSO_EVENTS order_by="venue_title"] returns an empty list.

In line 273 you're referencing with the alias EE_Venue_TBL, but that alias is never defined. I guess it's supposed to be "Venue", which is defined in the join in line 431.