do not assume offset is based on gmt_offset. [touch: 7056]
I realized that an assumption was being made here that the incoming timestamp would always have an offset applied in accordance with what is set as the sites timezone.  That's not necessarily true (although for current code in the wild it is very likely to be true).  So this simply corrects that assumption and makes the code more future proof.