fix _wp_http_referer piling up multiple copies of referers on each filter applied. [touch: 5685]
This is one of those things that isn't noticeable but if a user was to keep changing filters and applying them over and over again, eventually the url generated on page loads would reach the string limit and wierd things could happen.  Edgecase, but this fixes that potential bug.  Also means that the _wp_http_referer value in the request will actually be useful for "stickiness".