fix conditionals to be clearer [touch: 7901]
also take note that the conditional for the "FHEE__EED_Messages___maybe_registration__deliver_notifications" filter was implemented incorrectly I think.  The way it WAS setup indicated that IF that filter returned true messages were sent regardless of any other condition matches.  So we just need to have its value override any other condition matches if false.

Fixing this condition revealed another problem, in that (at least for the invoice payment method) the "finalize_registration" step does not appear to get initialized (or is skipped?) and thus messages never gets triggered.  I'm wondering if this is just a offline payment method thing.