fix queries
[touch: 7792]
[touch: 7056]

Note: if one is setting up a query where you ALREADY HAVE the date in the given timezone, then you should not send in the unixtimestamp equivalent because that is NOT treated as in the client timezone but is treated as UTC+0 (which is what a unix timestamp is).  So you need to send it in as a non 'U' format so it does NOT get converted.