switch to using day intervals.
month intervals for unit tests would fail because of the February conundrum and were too sensitive to WHEN the test was run.

I think this will still be too sensitive, if we get fails in the future I'm going to have to put in some more complicated interval calculation dependent on what "now" is when the test is run.

To be clear, this is *specifically* a unit test issue.