fix magically appearing ticket [touch: 8189]
Only add default ticket to datetime if this is a new event and it has not related tickets or if this is not a new event and there is only ONE datetime.

Note: this is still not the ideal fix because if the js fails for protecting against a ticket being orphaned in the dom, then this will not catch that scenario and its possible that there will still be no tickets attached to any datetimes.  however this is even more edgecase than the original scenario.  The ideal fix is to simply ALLOW users to have no tickets on a datetime but there is already a ticket for that.