ADDED filters to the Price and Quantity table headers.
This enables people to easily modify the content of these table headers and matches the pattern with the 'Available Tickets' header that already contains a filter.

Also, the event ID is passed to the filters. This enables, for example, a plugin/theme to be able to see if this event only allows people to select 1 ticket, in which case the radio button is displayed and in such a case the 'Qty' heading becomes superfluous. Same deal with 'Price' if the event is Free.