fix get_registrations query [touch: 8437]
The problem discovered by unit tests, is that we were calculating the days on a month based on the current date rather than the month sent in for the date range query.  So if the current month was June (which has 30 days) and the range query sent in July (which has 31 days) then the end date for the range would be 2015-06-31, which of course is invalid and gets converted to 2015-07-01 by the model method.