add filter for get_data_handler() [touch: 2195]
This is initially added for easier unit testing but also has utility for plugins that may want to swap out a data-handler on an existing message type for a custom one.  Or there may be certain messengers that require data in a different format than what the default data_handler's provide.