move code from EE_messages to EED_Messages [touch: 2195]
This is to keep the classes specific to their concerns.  EED_Messages handles all triggers and then passes off processing to EE_Messages_Processor.  EE_messages is now only acting sort of as a repository for messenger and message type info only.

Note: the methods are moved but the old locations are still present and act as wrappers for the new.  The old locations just have a @deprecated flag for now, but in future versions (likely beginning with 4.9.10 or 4.10.0), we'll throw a deprecated notice.