fix EEM_Payment_Test [touch: 8641]
the problem that was in this test is that there was a date getting created from `date( 'Y-m-d' )` which will return the Year, Month, Day for UTC time, and sent in AS America/New York time along with a specified time of 13:00:00.  So New York is roughly 7 hours behind UTC time, which meant that beginning around 5pm EST this test would have the date set one day ahead at 13:00:00 instead of the same day at 13:00:00.

The original change was made because of a though that $now wasn't reliable for tests, but the original problem was actually that $now was being manipulated without getting cloned (which I fixed recently but didn't realize it hadn't really fixed because of the issue described above).  Anyways, this test NOW passes fine.