move hook sets outside of config constructor [touch: 8799]
Unit tests demonstrate that if the hook is set in the constructor it will not have the correct values for the properties in teh callbacks.  Having the hooks set AFTER properties have been overridden from the db values results in proper values being used by callbacks.

So:
- if you want original default properties to be used in callback, set the hook in constructor.
- if you want actual properties (after retrieval from db), set the hook in the do_hooks() method.

I think the reason why this is needed is because the db is retrieving a serialized copy of config object and when it is unserialized it is setting the hook on that new unserialized object.  The other thing this helps prevent is hooks getting set multiple times.