add verison to deprecated tag [touch: 9199]
This is the usual standard for using the deprecated tag so its know WHEN it was deprecated (helps us also gauge when we can safely remove the method down the road (i.e. if we are at version 5.1 and it's been a year since we deprecated, likely would be safe and at that point we might not remember WHEN we deprecated this method).