remove unnecessary param [touch: 9281]
This was a work in progress but in the end its not needed because the incoming message types will be validated as installed.  If client code needs to validate _EXISTING_ message types it can call the validate_active_message_types_are_installed method (which is currently called by EEH_Activation).