fix logic for send_now check [touch: 9211]
EE_messenger objects return a simple boolean true/false for the value of the send now.  Whereas priority is derived from the EE_Message object which in turn is derived from the related EE_message_type and is the value for a EEM_Message priority constant.  These constants set the priority values as an int.

So basically the logic that was present was only doing a strict check for if the priority was equal to the int 10.  That covers for any message types that have a high priority but is NOT covering for when messenger's override the priority via send_now set to true.  So this commit fixes that.