put back in param that was removed [touch: 9211]
I think this was removed because the param wasn't actually used.  However, it was _intended_ to be used, it just didn't get implememented.  So I corrected that.

Basically its intended that the request key can be sent in, for any cases where the registration key gets obscufucated (it's just allowing for possible future changes.