only instantiate _cpt_model_object from this context if needed [touch: 9615]
This isn't directly related to the issue in this ticket, however this is a quick win for optimization to only re query the db if needed (although model caching should take care of it).  Basically it is possible for the _cpt_model_obj to not be already instantiated if this method is being hit from the context of the list table view actions as opposed to being hooked in on delete_post.