restore comparison of existing value in db vs current value in memory [touch: 9791]
The code must compare existing saved config with config in memory because if there is no difference that means that the method executed fine but there just was no update.  WordPress doesn't distinguish between false as a result 0 records updated because of no change vs false because some error produced problems with the update.