removing comparison that simply does not work [touch: 9872]
I don’t know why it works with single site and not in multisite, but in multisite, when doing an update, the value of $original was the same as the value of $this (even the properties!).  so the solution is to just always update regardless of whether there may be changes or not.  Since this is not used frequently there should be no harm.