make send_message final and catch exceptions [touch: 10057]
No child classes should be overriding this method because it has functionality that should be common to EVERY messenger.  That’s why this calls `_send_message` (which is to be implemented by child classes).  Hence the change to make this final.

This then allows us to take care of preparing a consistent response for client code to handle the response from a send attempt.

In here we also then implement the new `SendMessageException` and improve phpdoc blocks.