Fix usage of $dttrow [touch: 10292]
For some reason, even though `DTT_order` was set on `EE_Datetime` objects earlier in the request, calling `$dtt->get( ‘DTT_order’ )` was returning the original value on the datetime.  This meant that if all datetimes have the same value (which can happen), then the datetime-list items prepared for tickets would be the same active items for all tickets (which is incorrect).

Since we’re ALREADY passing in the correct value for `$dttrow` into the method, let’s just use that.