add a new static property on EE_Registry that is used to hold any data that needs exposed in the dom for javascript loading on a request. [touch: 10304]
Note:

Initially this will get enqueues on the `eejs.data` object and attached to `eejs-api`.  The problem with that is that eventually I’d lke to move `EE_Registry::i18n_js_strings` onto here as well.

Also, long term, ideally I think it’d be better if we have a `Scripts` class in the `EventEspresso` namespace.  And that would get loaded via dependency injection on a `DisplayRequest` request type whenever our request handling in EE gets refactored.  However, I didn’t want to get distracted by implementing that now so I’ve just added it this way for now.