use the new `EE_Registry::$eejs` and register it with `eejs-api` [touch: 10304]
Again, not ideal that this is attached to `eejs-api` because its more of a global type property that we want accessible when ANY javascript we build that uses eejs is loaded. So, eventually (maybe when `espresso_core.js` gets a refactor) when we have a base.js file that would become a dependency on ANY js file in our system, we could attach it to that.