Maybe override default setting for EE_Messages_Config::keep_messages_forever config [touch: 10417]
This check is only done once and then if the site has existing messages past the default threshold then the default will be changed from false to true, so that the sites messages will be kept forever.  This is because we don’t want to _assume_ that ALL users will want their old messages deleted if there are already messages in the db.