modify method to account for changes with cleanup options [touch: 10417]
Since this is a public method, I decided to not have this dependent on the config option but instead receive a parameter for what the threshold is.  Also I left the filter in because client code may want the incoming parameter to ignored everytime this method is called and only a specific value used.