add handling for _links and _calculated_fields property from an entity response [touch: 10454]
Some other fixes in here as well.

Separating these things out still exposes them for use by client code but keeps them distinct from the entity stored in the state (and they shouldn’t be mutable).  Eventually we could use the _links for grabbing what endpoint to use for REST request, and we could expose calculated fields as a computed property on the model components.