consider that the first timezone_string match to a given offset could be invalid [touch: 10626]
The php methods, `timezone_name_from_abbr` and `timezone_abbreviations_list` are known to not only hold references to _current_ timezone offset records but also _historical_ timezone offset records.   However, there is no indicator within these results what is historical and what is current.  This means that the first match to a given offset might be correct historically but NOT correct currently.

So what this change does, is use the match to generated a new timezone object and get the _current_ offset returned by PHP for that timezone.  If it doesn’t match the offset supplied for checking, then we discard the match and continue.