only test valid offsets and conduct proper assertion for UTC +0 offset [touch: 10626]
The problem with testing invalid offsets is that WordPress uses gmt offsets AS is without attempting to coerce to a valid timezone_string.  Since `current_time()` is being used in this test, that means that there will always be a conflict with the times saved by EE for invalid offsets which are coerced to valid timezones for use internally by the EE system.